system-upgrade: Add an env variable to control reboot
If the DNF_SYSTEM_UPGRADE_NO_REBOOT env variable is set, the plugin will
not actually reboot.

= changelog =
msg: Add the DNF_SYSTEM_UPGRADE_NO_REBOOT env variable to control system-upgrade reboot.
type: enhancement

+++ plugins/system_upgrade.py
@@ -71,7 +71,10 @@ STATE_VERSION = 2
 
 
 def reboot():
-    Popen(["systemctl", "reboot"])
+    if os.getenv("DNF_SYSTEM_UPGRADE_NO_REBOOT", default=False):
+        logger.info(_("Reboot turned off, not rebooting."))
+    else:
+        Popen(["systemctl", "reboot"])
 
 
 def get_url_from_os_release():
